Convert sets tests to nested test
